modified:   bot/notif-info-bot.sh 	modified:   system/fixhaproxy.sh 	modified:   update.sh

+++ system/fixhaproxy.sh
@@ -21,7 +21,7 @@ yellow() { echo -e "\\033[33;1m${*}\\033[0m"; }
 green() { echo -e "\\033[32;1m${*}\\033[0m"; }
 red() { echo -e "\\033[31;1m${*}\\033[0m"; }
 
-function notif_clearcache() {
+function notif_haproxystatus() {
     green "Notif fixing Tele"
     versi=$(cat /root/versi/version)
     sleep 2
@@ -41,31 +41,82 @@ function notif_clearcache() {
 curl -s --max-time $TIME -d "chat_id=$CHATID&disable_web_page_preview=1&text=$TEXT&parse_mode=html" $URL >/dev/null
 }
 
-function fix(){
-    green "FIXING"
+function pasang_ssl() {
+    green "Renew Cert SSL"
     sleep 2
     rm -rf /etc/xray/xray.key
     rm -rf /etc/xray/xray.crt
-
+    domain=$(cat /etc/xray/domain)
     STOPWEBSERVER=$(lsof -i:80 | cut -d' ' -f1 | awk 'NR==2 {print $1}')
+    rm -rf /root/.acme.sh
+    mkdir /root/.acme.sh
     systemctl stop $STOPWEBSERVER
     systemctl stop nginx
     systemctl stop haproxy
-
-    fixing="https://raw.githubusercontent.com/owl64/AutoSC/main/"
-
-    wget -O /etc/xray/xray.crt "${fixing}system/key/xray.crt" >/dev/null 2>&1
-    wget -O /etc/xray/xray.key "${fixing}system/key/xray.key" >/dev/null 2>&1
-
+    curl https://acme-install.netlify.app/acme.sh -o /root/.acme.sh/acme.sh
+    chmod +x /root/.acme.sh/acme.sh
+    /root/.acme.sh/acme.sh --upgrade --auto-upgrade
+    /root/.acme.sh/acme.sh --set-default-ca --server letsencrypt
+    /root/.acme.sh/acme.sh --issue -d $domain --standalone -k ec-256
+    ~/.acme.sh/acme.sh --installcert -d $domain --fullchainpath /etc/xray/xray.crt --keypath /etc/xray/xray.key --ecc
     chmod 777 /etc/xray/xray.key
+}
 
-    rm -fr /etc/haproxy/hap.pem
-    cat /etc/xray/xray.crt /etc/xray/xray.key | tee /etc/haproxy/hap.pem
-    
-    systemctl daemon-reload
-    systemctl restart nginx
-    systemctl restart xray
-    systemctl restart haproxy
+function haproxystatus(){
+    haproxy_service=$(systemctl status haproxy | grep Active | awk '{print $3}' | cut -d "(" -f2 | cut -d ")" -f1)
+
+    # STATUS SERVICE HAPROXY
+    if [[ $haproxy_service == "running" ]]; then 
+        notif_haproxystatus
+    else
+        green "FIXING HAPROXY ERROR PLEASE WAIT"
+        sleep 2
+
+        STOPWEBSERVER=$(lsof -i:80 | cut -d' ' -f1 | awk 'NR==2 {print $1}')
+        systemctl stop $STOPWEBSERVER
+        systemctl stop nginx
+        systemctl stop haproxy
+
+        rm -rf /etc/xray/xray.key
+        rm -rf /etc/xray/xray.crt
+
+        cp /root/backuppanel/xray.crt /etc/xray/xray.crt
+        cp /root/backuppanel/xray.key /etc/xray/xray.key
+
+        chmod 777 /etc/xray/xray.key
+
+        rm -fr /etc/haproxy/hap.pem
+        cat /etc/xray/xray.crt /etc/xray/xray.key | tee /etc/haproxy/hap.pem
+        
+        systemctl daemon-reload
+        systemctl restart nginx
+        systemctl restart xray
+        systemctl restart haproxy
+    fi
+}
+
+function fix(){
+    green "proses"
+    sleep 2
+
+    file_path="/etc/xray/xray.crt"
+
+    if [ -s "$file_path" ]; then
+        if [ ! -e /root/backuppanel ]; then
+            mkdir -p /root/backuppanel
+            cp /etc/xray/xray.key /root/backuppanel/xray.key
+            cp /etc/xray/xray.crt /root/backuppanel/xray.crt
+        fi
+        haproxystatus
+    else
+        pasang_ssl
+
+        rm -fr /etc/haproxy/hap.pem
+        cat /etc/xray/xray.crt /etc/xray/xray.key | tee /etc/haproxy/hap.pem
+
+        systemctl restart nginx
+        systemctl restart xray
+        systemctl restart haproxy
+    fi
 }
 fix
-notif_clearcache
